+++ packages/generator/lib/combo/cosmetics/index.ts
@@ -430,7 +430,7 @@ class CosmeticsPass {
       this.patchSymbol('MUSIC_CUSTOM', new Uint8Array([0x01]));
       const data = await this.getPathBuffer(c.music);
       if (data)
-        await randomizeMusic(this.logWriter, this.monitor, this.builder, random, data);
+        await randomizeMusic(c, this.logWriter, this.monitor, this.builder, random, data);
     }
     if (c.musicNames) {
       this.patchSymbol('MUSIC_NAMES', new Uint8Array([0x01]));